Move flush-bug test to test/append.js
Problem: The `flush-bug.js` file has only one test and would be better
integrated with another file.

Solution: Move the test to `test/append.js`, which has other similar
tests.
